Refactor Skills and move shared logic into SkillButton
